split out mention resolver, remove blocking thread usage
